+++ CHANGELOG.md
@@ -15,6 +15,7 @@ You can find the changelogs for the individual packages in their respective `CHA
 - [@ai-sdk/deepseek](./packages/deepseek/CHANGELOG.md)
 - [@ai-sdk/fal](./packages/fal/CHANGELOG.md)
 - [@ai-sdk/fireworks](./packages/fireworks/CHANGELOG.md)
+- [@ai-sdk/gateway](./packages/gateway/CHANGELOG.md)
 - [@ai-sdk/google](./packages/google/CHANGELOG.md)
 - [@ai-sdk/google-vertex](./packages/google-vertex/CHANGELOG.md)
 - [@ai-sdk/groq](./packages/groq/CHANGELOG.md)